+++ src/test/java/my/store/qa/locators/MyStoreLocators/MyStoreMainPageLocators.java
@@ -1,0 +1,19 @@
+package my.store.qa.locators.MyStoreLocators;
+
+import org.openqa.selenium.By;
+
+public class MyStoreMainPageLocators {
+
+    private static By LOGO_MY_STORE;
+
+    public MyStoreMainPageLocators() {
+
+        LOGO_MY_STORE = By.cssSelector("[id=\"header_logo\"]");
+    }
+
+    public By getLogoLocator() {
+
+        return LOGO_MY_STORE;
+    }
+
+}
