+++ src/test/java/my/store/qa/pages/MyStorePages/MyStoreMainPage.java
@@ -1,0 +1,30 @@
+package my.store.qa.pages.MyStorePages;
+
+import my.store.qa.locators.MyStoreLocators.MyStoreMainPageLocators;
+import my.store.qa.pages.BasePage;
+import my.store.qa.world.Environment;
+
+import java.time.Duration;
+
+public class MyStoreMainPage extends BasePage {
+
+    private final MyStoreMainPageLocators locators;
+
+    public MyStoreMainPage() {
+
+        locators = new MyStoreMainPageLocators();
+    }
+
+    public void openMyStore() {
+
+        String url = Environment.getMyStoreConfiguration().getMyStorePath();
+        open(url);
+
+    }
+
+    public boolean isMyStoreLogoDisplayed() {
+
+        return isElementDisplayedByLocator(locators.getLogoLocator(), Duration.ofSeconds(1));
+    }
+
+}
